+++ include/xtensor/xchunked_view.hpp
@@ -10,141 +10,125 @@
 #ifndef XTENSOR_CHUNKED_VIEW_HPP
 #define XTENSOR_CHUNKED_VIEW_HPP
 
-#include "xstrided_view.hpp"
+#include <xtl/xsequence.hpp>
+
 #include "xnoalias.hpp"
+#include "xstorage.hpp"
+#include "xstrided_view.hpp"
+#include "xchunked_array.hpp"
 
 namespace xt
 {
 
     template <class E>
-    class xchunked_view;
+    struct is_chunked_t: detail::chunk_helper<E>::is_chunked
+    {
+    };
+
+    /*****************
+     * xchunked_view *
+     *****************/
 
     template <class E>
-    class xchunk_iterator
-    {
-    public:
-        xchunk_iterator(xchunked_view<E>& chunked_view, std::size_t chunk_idx);
-        xchunk_iterator() = default;
-
-        xchunk_iterator<E>& operator++();
-        xchunk_iterator<E> operator++(int);
-        bool operator==(const xchunk_iterator& other) const;
-        bool operator!=(const xchunk_iterator& other) const;
-        auto operator*();
-
-    private:
-        xchunked_view<E>* m_pcv;
-        std::size_t m_ci;
-    };
+    class xchunk_iterator;
 
     template <class E>
     class xchunked_view
     {
     public:
-        template <class OE>
-        xchunked_view(OE&& e, std::vector<std::size_t>& chunk_shape);
 
-        xchunk_iterator<E> begin();
-        xchunk_iterator<E> end();
+        using self_type = xchunked_view<E>;
+        using expression_type = std::decay_t<E>;
+        using value_type = typename expression_type::value_type;
+        using reference = typename expression_type::reference;
+        using const_reference = typename expression_type::const_reference;
+        using pointer = typename expression_type::pointer;
+        using const_pointer = typename expression_type::const_pointer;
+        using size_type = typename expression_type::size_type;
+        using difference_type = typename expression_type::difference_type;
+        using shape_type = svector<size_type>;
+        using chunk_iterator = xchunk_iterator<self_type>;
+        using const_chunk_iterator = xchunk_iterator<const self_type>;
+
+        template <class OE, class S>
+        xchunked_view(OE&& e, S&& chunk_shape);
 
         template <class OE>
-        xchunked_view<E>& operator=(const OE& e);
+        xchunked_view(OE&& e);
+
+        void init();
+
+        template <class OE>
+        typename std::enable_if_t<!is_chunked_t<OE>::value, xchunked_view<E>&> operator=(const OE& e);
+
+        template <class OE>
+        typename std::enable_if_t<is_chunked_t<OE>::value, xchunked_view<E>&> operator=(const OE& e);
+
+        size_type dimension() const noexcept;
+        const shape_type& shape() const noexcept;
+        const shape_type& chunk_shape() const noexcept;
+        size_type grid_size() const noexcept;
+        const shape_type& grid_shape() const noexcept;
+
+        expression_type& expression() noexcept;
+        const expression_type& expression() const noexcept;
+
+        chunk_iterator chunk_begin();
+        chunk_iterator chunk_end();
+
+        const_chunk_iterator chunk_begin() const;
+        const_chunk_iterator chunk_end() const;
+        const_chunk_iterator chunk_cbegin() const;
+        const_chunk_iterator chunk_cend() const;
 
     private:
-        E m_expression;
-        std::vector<std::size_t> m_shape;
-        std::vector<std::size_t> m_chunk_shape;
-        std::vector<std::size_t> m_shape_of_chunks;
-        std::vector<std::size_t> m_ic;
-        std::size_t m_chunk_nb;
-        xstrided_slice_vector m_sv;
 
-        friend class xchunk_iterator<E>;
+        E m_expression;
+        shape_type m_shape;
+        shape_type m_chunk_shape;
+        shape_type m_grid_shape;
+        size_type m_chunk_nb;
     };
 
-    template <class E>
-    inline xchunk_iterator<E>::xchunk_iterator(xchunked_view<E>& chunked_view, std::size_t chunk_idx)
-        : m_pcv(&chunked_view)
-        , m_ci(chunk_idx)
-    {
-    }
+    template <class E, class S>
+    xchunked_view<E> as_chunked(E&& e, S&& chunk_shape);
+
+    /********************************
+     * xchunked_view implementation *
+     ********************************/
 
     template <class E>
-    inline xchunk_iterator<E>& xchunk_iterator<E>::operator++()
-    {
-        if (m_ci != m_pcv->m_chunk_nb - 1)
-        {
-            std::size_t di = m_pcv->m_shape.size() - 1;
-            while (true)
-            {
-                if (m_pcv->m_ic[di] + 1 == m_pcv->m_shape_of_chunks[di])
-                {
-                    m_pcv->m_ic[di] = 0;
-                    m_pcv->m_sv[di] = range(0, m_pcv->m_chunk_shape[di]);
-                    if (di == 0)
-                    {
-                        break;
-                    }
-                    else
-                    {
-                        di--;
-                    }
-                }
-                else
-                {
-                    m_pcv->m_ic[di] += 1;
-                    m_pcv->m_sv[di] = range(m_pcv->m_ic[di] * m_pcv->m_chunk_shape[di], (m_pcv->m_ic[di] + 1) * m_pcv->m_chunk_shape[di]);
-                    break;
-                }
-            }
-        }
-        m_ci++;
-        return *this;
-    }
-
-    template <class E>
-    inline xchunk_iterator<E> xchunk_iterator<E>::operator++(int)
-    {
-        xchunk_iterator<E> it = *this;
-        ++(*this);
-        return it;
-    }
-
-    template <class E>
-    inline bool xchunk_iterator<E>::operator==(const xchunk_iterator& other) const
-    {
-        return m_ci == other.m_ci;
-    }
-
-    template <class E>
-    inline bool xchunk_iterator<E>::operator!=(const xchunk_iterator& other) const
-    {
-        return !(*this == other);
-    }
-
-    template <class E>
-    inline auto xchunk_iterator<E>::operator*()
-    {
-        auto chunk = strided_view(m_pcv->m_expression, m_pcv->m_sv);
-        return std::make_pair(chunk, m_pcv->m_sv);
-    }
-
-    template <class E>
-    template <class OE>
-    inline xchunked_view<E>::xchunked_view(OE&& e, std::vector<std::size_t>& chunk_shape)
+    template <class OE, class S>
+    inline xchunked_view<E>::xchunked_view(OE&& e, S&& chunk_shape)
         : m_expression(std::forward<OE>(e))
-        , m_chunk_shape(chunk_shape)
+        , m_chunk_shape(xtl::forward_sequence<shape_type, S>(chunk_shape))
     {
         m_shape.resize(e.dimension());
         const auto& s = e.shape();
         std::copy(s.cbegin(), s.cend(), m_shape.begin());
+        init();
+    }
+
+    template <class E>
+    template <class OE>
+    inline xchunked_view<E>::xchunked_view(OE&& e)
+        : m_expression(std::forward<OE>(e))
+    {
+        m_shape.resize(e.dimension());
+        const auto& s = e.shape();
+        std::copy(s.cbegin(), s.cend(), m_shape.begin());
+    }
+
+    template <class E>
+    void xchunked_view<E>::init()
+    {
         // compute chunk number in each dimension
-        m_shape_of_chunks.resize(m_shape.size());
+        m_grid_shape.resize(m_shape.size());
         std::transform
         (
             m_shape.cbegin(), m_shape.cend(),
             m_chunk_shape.cbegin(),
-            m_shape_of_chunks.begin(),
+            m_grid_shape.begin(),
             [](auto s, auto cs)
             {
                 std::size_t cn = s / cs;
@@ -155,46 +139,143 @@ namespace xt
                 return cn;
             }
         );
-        m_ic.resize(m_shape.size());
-        m_chunk_nb = std::accumulate(std::begin(m_shape_of_chunks), std::end(m_shape_of_chunks), std::size_t(1), std::multiplies<>());
-        m_sv.resize(m_shape.size());
-    }
-
-    template <class E>
-    inline xchunk_iterator<E> xchunked_view<E>::begin()
-    {
-        auto it = xchunk_iterator<E>(*this, 0);
-        std::transform(m_chunk_shape.begin(), m_chunk_shape.end(), m_sv.begin(),
-                       [](auto size) { return range(0, size); });
-        std::fill(m_ic.begin(), m_ic.end(), std::size_t(0));
-        return it;
-    }
-
-    template <class E>
-    inline xchunk_iterator<E> xchunked_view<E>::end()
-    {
-        auto it = xchunk_iterator<E>(*this, m_chunk_nb);
-        return it;
+        m_chunk_nb = std::accumulate(std::begin(m_grid_shape), std::end(m_grid_shape), std::size_t(1), std::multiplies<>());
     }
 
     template <class E>
     template <class OE>
-    xchunked_view<E>& xchunked_view<E>::operator=(const OE& e)
+    typename std::enable_if_t<!is_chunked_t<OE>::value, xchunked_view<E>&> xchunked_view<E>::operator=(const OE& e)
     {
-        for (auto it = begin(); it != end(); it++)
+        auto end = chunk_end();
+        for (auto it = chunk_begin(); it != end; ++it)
         {
             auto el = *it;
-            noalias(el.first) = strided_view(e, el.second);
+            noalias(el) = strided_view(e, it.get_slice_vector());
         }
         return *this;
     }
 
     template <class E>
-    auto as_chunked(E&& e, std::vector<std::size_t>& chunk_shape)
+    template <class OE>
+    typename std::enable_if_t<is_chunked_t<OE>::value, xchunked_view<E>&> xchunked_view<E>::operator=(const OE& e)
     {
-        return xchunked_view<E>(std::forward<E>(e), chunk_shape);
+        m_chunk_shape.resize(e.dimension());
+        const auto& cs = e.chunk_shape();
+        std::copy(cs.cbegin(), cs.cend(), m_chunk_shape.begin());
+        init();
+        auto it2 = e.chunks().begin();
+        auto end1 = chunk_end();
+        for (auto it1 = chunk_begin(); it1 != end1; ++it1, ++it2)
+        {
+            auto el1 = *it1;
+            auto el2 = *it2;
+            auto lhs_shape = el1.shape();
+            if (lhs_shape != el2.shape())
+            {
+                xstrided_slice_vector esv(el2.dimension());  // element slice in edge chunk
+                std::transform(lhs_shape.begin(), lhs_shape.end(), esv.begin(),
+                               [](auto size) { return range(0, size); });
+                noalias(el1) = strided_view(el2, esv);
+            }
+            else
+            {
+                noalias(el1) = el2;
+            }
+        }
+        return *this;
     }
 
+    template <class E>
+    inline auto xchunked_view<E>::dimension() const noexcept -> size_type
+    {
+        return m_shape.size();
+    }
+
+    template <class E>
+    inline auto xchunked_view<E>::shape() const noexcept -> const shape_type&
+    {
+        return m_shape;
+    }
+
+    template <class E>
+    inline auto xchunked_view<E>::chunk_shape() const noexcept -> const shape_type&
+    {
+        return m_chunk_shape;
+    }
+
+    template <class E>
+    inline auto xchunked_view<E>::grid_size() const noexcept -> size_type
+    {
+        return m_chunk_nb;
+    }
+
+    template <class E>
+    inline auto xchunked_view<E>::grid_shape() const noexcept -> const shape_type&
+    {
+        return m_grid_shape;
+    }
+
+    template <class E>
+    inline auto xchunked_view<E>::expression() noexcept -> expression_type&
+    {
+        return m_expression;
+    }
+
+    template <class E>
+    inline auto xchunked_view<E>::expression() const noexcept -> const expression_type&
+    {
+        return m_expression;
+    }
+
+    template <class E>
+    inline auto xchunked_view<E>::chunk_begin() -> chunk_iterator
+    {
+        shape_type chunk_index(m_shape.size(), size_type(0));
+        return chunk_iterator(*this, std::move(chunk_index), 0u);
+    }
+
+    template <class E>
+    inline auto xchunked_view<E>::chunk_end() -> chunk_iterator
+    {
+        return chunk_iterator(*this, shape_type(grid_shape()), grid_size());
+    }
+
+    template <class E>
+    inline auto xchunked_view<E>::chunk_begin() const -> const_chunk_iterator
+    {
+        shape_type chunk_index(m_shape.size(), size_type(0));
+        return const_chunk_iterator(*this, std::move(chunk_index), 0u);
+    }
+
+    template <class E>
+    inline auto xchunked_view<E>::chunk_end() const -> const_chunk_iterator
+    {
+        return const_chunk_iterator(*this, shape_type(grid_shape()), grid_size());
+    }
+
+    template <class E>
+    inline auto xchunked_view<E>::chunk_cbegin() const -> const_chunk_iterator
+    {
+        return chunk_begin();
+    }
+
+    template <class E>
+    inline auto xchunked_view<E>::chunk_cend() const -> const_chunk_iterator
+    {
+        return chunk_end();
+    }
+
+    template <class E, class S>
+    inline xchunked_view<E> as_chunked(E&& e, S&& chunk_shape)
+    {
+        return xchunked_view<E>(std::forward<E>(e), std::forward<S>(chunk_shape));
+    }
+
+    template <class E>
+    inline xchunked_view<E> as_chunked(E&& e)
+    {
+        return xchunked_view<E>(std::forward<E>(e));
+    }
 }
 
 #endif